+++ src/netspresso_trainer/models/heads/classification/__init__.py
@@ -1,1 +1,1 @@
-from .experimental.fc import fc
+from .core.fc import fc
